# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da API do Press Ticket®",
    "description": "Documentação da API para envio de mensagens e gerenciamento de contatos.",
    "contact": {
      "email": "robson.tenorio@gmail.com"
    },
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "https://devapi.pressticket.com.br",
      "description": "API Press Ticket®"
    }
  ],
  "paths": {
...
    "/v1/tags": {
      "get": {
        "summary": "Listar Tags",
        "description": "Listar todas as tags via API\n\nPermissões requeridas: `read:tags`",
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "apiToken": [
              "read:tags"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tags"
          },
          "401": {
            "description": "Token inválido ou não fornecido"
          },
          "403": {
            "description": "Token não tem permissão 'read:tags'"
          },
          "500": {
            "description": "Erro interno!"
          }
        }
      },
      "post": {
        "summary": "Criar Tag",
        "description": "Criar uma nova tag via API\n\nPermissões requeridas: `create:tags`",
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "apiToken": [
              "create:tags"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nome da tag"
                  },
                  "color": {
                    "type": "string",
                    "description": "Cor da tag em formato hexadecimal (ex: #0000FF)"
                  }
                },
                "required": [
                  "name",
                  "color"
                ]
              },
              "examples": {
                "tag": {
                  "value": {
                    "name": "Tag de Teste API",
                    "color": "#0000FF"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag criada com sucesso"
          },
          "401": {
            "description": "Token inválido ou não fornecido"
          },
          "403": {
            "description": "Token não tem permissão 'create:tags'"
          },
          "500": {
            "description": "Erro interno!"
          }
        }
      }
    },
...
    "/v1/tags/{tagId}": {
      "get": {
        "summary": "Obter Tag",
        "description": "Obter detalhes de uma tag específica via API\n\nPermissões requeridas: `read:tags`",
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "apiToken": [
              "read:tags"
            ]
          }
        ],
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "description": "ID da tag",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag encontrada"
          },
          "401": {
            "description": "Token inválido ou não fornecido"
          },
          "403": {
            "description": "Token não tem permissão 'read:tags'"
          },
          "404": {
            "description": "Tag não encontrada"
          },
          "500": {
            "description": "Erro interno!"
          }
        }
      },
      "put": {
        "summary": "Atualizar Tag",
        "description": "Atualizar uma tag específica via API\n\nPermissões requeridas: `update:tags`",
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "apiToken": [
              "update:tags"
            ]
          }
        ],
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "description": "ID da tag",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nome da tag"
                  },
                  "color": {
                    "type": "string",
                    "description": "Cor da tag em formato hexadecimal (ex: #00FFFF)"
                  }
                }
              },
              "examples": {
                "tag": {
                  "value": {
                    "name": "Tag Atualizada API",
                    "color": "#00FFFF"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag atualizada com sucesso"
          },
          "401": {
            "description": "Token inválido ou não fornecido"
          },
          "403": {
            "description": "Token não tem permissão 'update:tags'"
          },
          "404": {
            "description": "Tag não encontrada"
          },
          "500": {
            "description": "Erro interno!"
          }
        }
      },
      "delete": {
        "summary": "Excluir Tag",
        "description": "Excluir uma tag específica via API\n\nPermissões requeridas: `delete:tags`",
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "apiToken": [
              "delete:tags"
            ]
          }
        ],
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "description": "ID da tag",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag excluída com sucesso"
          },
          "401": {
            "description": "Token inválido ou não fornecido"
          },
          "403": {
            "description": "Token não tem permissão 'delete:tags'"
          },
          "404": {
            "description": "Tag não encontrada"
          },
          "500": {
            "description": "Erro interno!"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "apiToken": {
        "type": "apiKey",
        "name": "x-api-token",
        "in": "header",
        "description": "Token de API para autenticação. Requer permissões específicas para cada rota:\n- `create:messages`: Para enviar mensagens\n- `create:contacts`: Para criar contatos"
      }
    }
  }
}